LIBFCREPO-1034. Renamed plastron.commands.import to plastron.commands.importcommand.
Simplified the import command tests.

https://issues.umd.edu/browse/LIBFCREPO-1034
